Improve usage of OIDC_TLS_VERIFY (#78)
* Fix inverted logic of OIDC_TLS_VERIFY

When OIDC_TLS_VERIFY was set to the string "true" this was converted
to the boolean `true` by `strToBool`. This resulted in `skipTLSVerify`
also returning `true`. Thus verification was actually skipped.

Fixed this inverted logic bug.

* Apply OIDC_TLS_VERIFY to OIDC provider as well

The configuration OIDC_TLS_VERIFY was only applied to for the refresh-
token related checks. It was still not possible to use the test client
with an OIDC provider that was using a self-signed certificate.

This commit changes the context used for communicating with the OIDC
provider. That way TLS certificate validation can be skipped and thus
an OIDC provider using a self-signed certificate can also be used.